requirements.txt
amqp==2.5.2
asgiref==3.2.7
astroid==2.4.1
billiard==3.6.3.0
boto==2.49.0
boto3==1.13.19
botocore==1.16.19
celery==4.4.2
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
decorator==4.4.2
Django==2.2.6
django-cors-headers==3.1.1
django-rest-swagger==2.1.1
django-storages==1.9.1
djangorestframework==3.9.4
docutils==0.15.2
idna==2.9
imageio==2.8.0
imageio-ffmpeg==0.4.2
importlib-metadata==1.6.0
inflection==0.4.0
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
kombu==4.6.8
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
moviepy==1.0.3
numpy==1.18.4
openapi-codec==1.3.2
packaging==20.4
Pillow==7.1.2
proglog==0.1.9
psycopg2==2.8.5
pylint==2.5.2
pymediainfo==4.2.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
simplejson==3.17.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
tqdm==4.46.0
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
wrapt==1.12.1
zipp==3.1.0
